Updated base exclusions class logic
